+++ src/dev/messages_de.js
@@ -1,8 +1,0 @@
-exports.messages = {
-  admin: 'Starte VS Code als Administrator um die Änderungen durchführen zu können.',
-  enabled: 'Icons aktiviert, starte VS Code bitte neu.',
-  disabled: 'Icons deaktiviert, starte VS Code bitte neu.',
-  already_disabled: 'Icons wurden bereits deaktiviert.',
-  smthingwrong: 'Ups! Etwas ist schief gelaufen: ',
-  restartIde: 'Visual Studio Code neu starten'
-};