+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Change Log
 
 ## 2.0.6 (September 24, 2016)
+  - Icons Enable not supported anymore for VSCode >= v1.6.0. ([@robertohuertasm](https://github.com/robertohuertasm) in [#353](https://github.com/robertohuertasm/vscode-icons/pull/353))
   - Enhancement: docs folder support. ([@robertohuertasm](https://github.com/robertohuertasm) in [#352](https://github.com/robertohuertasm/vscode-icons/pull/352))
   - Enhancement: xliff icon reduced. ([@robertohuertasm](https://github.com/robertohuertasm) in [#351](https://github.com/robertohuertasm/vscode-icons/pull/351))
   - Enhancement: support for images folder. ([@robertohuertasm](https://github.com/robertohuertasm) in [#350](https://github.com/robertohuertasm/vscode-icons/pull/350))